requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.8.6
requests==2.25.0
pytest==6.1.2
pytest-cov==2.10.1
python-dotenv==0.15.0